+++ src/background/keyboard-shortcut-handler.ts
